Ydicsey to main

+++ HealPlant.cpp
@@ -1,6 +1,6 @@
-#include <HealPlant.h>
+#include "HealPlant.h"
 
-HealPlant::HealPlant(const size_t& cost, const size_t& hp, const size_t& healPoint) : Plant(cost, hp), healPoint_(healPoint) {}
+HealPlant::HealPlant(const int& cost, const int& hp, const int& healPoint) : Plant(cost, hp), healPoint_(healPoint) {}
 
 const int& HealPlant::myHealPoint() const
 {
